Create sum_integr_digit.c
Basic working code statements. checked on the online compiler. which will give the sum of given integer digit.
This code logic is with the goto keyword, just for the reference.

Note : The integer value would be Limited here, used unsigned int and have not added any check for range value.